requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
slowapi>=0.1.9
streamlit>=1.36.0
plotly>=5.22.0
groq>=0.9.0
python-dotenv>=1.0.1
chromadb>=0.4.0
